refactor: move OptionSource and EloquentOptions to Core
They are now shared across Forms and Tables, so the contract and its Eloquent
adapter belong in Core (alongside Option) rather than Forms. The arch guards
forbid Tables depending on Forms, so this is what lets table and column filters
reuse the same option source the Select field uses.
