+++ workbench/app/Forms/ProductForm.php
@@ -8,11 +8,11 @@ use Illuminate\Validation\Rule;
 use Lattice\Lattice\Attributes\Form;
 use Lattice\Lattice\Core\Components\Card;
 use Lattice\Lattice\Core\Components\Grid;
+use Lattice\Lattice\Core\EloquentOptions;
 use Lattice\Lattice\Forms\Components\Choice;
 use Lattice\Lattice\Forms\Components\Form as FormComponent;
 use Lattice\Lattice\Forms\Components\Select;
 use Lattice\Lattice\Forms\Components\TextInput;
-use Lattice\Lattice\Forms\EloquentOptions;
 use Lattice\Lattice\Forms\FormDefinition;
 use Symfony\Component\HttpFoundation\Response;
 use Workbench\App\Models\Product;
